signature_docker/requirements.txt
numpy<2
opencv-python-headless
Pillow
